requirements.txt
flask
requests
python-binance
pandas
